# docs/contracts/local-api.openapi.yml
openapi: 3.0.3
info:
  title: Sistema MNSS Local API
  version: 0.0.1
  description: |
    API para operação local da Padaria Nova Aliança.
    
    ### WebSocket Events
    O sistema publica eventos em tempo real para sincronização de telas:
    
    - `/topic/kds/tickets`: Enviado quando um ticket é criado ou atualizado. Payload: `KdsTicketResponse`.
    - `/topic/kds/tickets/{sector}`: Mesmo evento acima, mas filtrado por setor (ex: `COZINHA`, `FORNO`).
    - `/topic/orders/ready`: Enviado quando um pedido completo está pronto para entrega. Payload: `UUID` (ID do pedido).
paths:
...
  /api/pdv/sales:
    get:
      summary: Lista vendas presenciais do PDV.
      operationId: listPdvSales
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Vendas PDV encontradas.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PdvSaleResponse'
        '403':
          $ref: '#/components/responses/BusinessError'
        default:
          $ref: '#/components/responses/UnexpectedError'
    post:
      summary: Inicia venda presencial no PDV.
      operationId: createPdvSale
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Venda criada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdvSaleResponse'
        '400':
          $ref: '#/components/responses/BusinessError'
        '403':
          $ref: '#/components/responses/BusinessError'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /api/pdv/sales/{saleId}:
    get:
      summary: Retorna venda presencial do PDV.
      operationId: getPdvSale
      security:
        - bearerAuth: []
      parameters:
        - name: saleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Venda encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdvSaleResponse'
        '404':
          $ref: '#/components/responses/BusinessError'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /api/pdv/sales/{saleId}/items:
    post:
      summary: Adiciona produto ao carrinho da venda PDV.
      operationId: addPdvSaleItem
      security:
        - bearerAuth: []
      parameters:
        - name: saleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePdvSaleItemRequest'
      responses:
        '201':
          description: Carrinho atualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdvSaleResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/BusinessError'
        '404':
          $ref: '#/components/responses/BusinessError'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /api/pdv/sales/{saleId}/items/{itemId}:
    patch:
      summary: Altera quantidade de um item da venda PDV.
      operationId: updatePdvSaleItem
      security:
        - bearerAuth: []
      parameters:
        - name: saleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchPdvSaleItemRequest'
      responses:
        '200':
          description: Carrinho atualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdvSaleResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/BusinessError'
        '404':
          $ref: '#/components/responses/BusinessError'
        default:
          $ref: '#/components/responses/UnexpectedError'
    delete:
      summary: Remove item do carrinho da venda PDV.
      operationId: removePdvSaleItem
      security:
        - bearerAuth: []
      parameters:
        - name: saleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Carrinho atualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdvSaleResponse'
        '403':
          $ref: '#/components/responses/BusinessError'
        '404':
          $ref: '#/components/responses/BusinessError'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /api/pdv/sales/{saleId}/payment:
    post:
      summary: Registra pagamento parcial ou total de uma venda PDV.
      operationId: payPdvSale
      security:
        - bearerAuth: []
      parameters:
        - name: saleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '201':
          description: Pagamento registrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/BusinessError'
        '404':
          $ref: '#/components/responses/BusinessError'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /api/pdv/sales/{saleId}/finish:
    post:
      summary: Finaliza a venda PDV após pagamentos completos.
      operationId: finishPdvSale
      security:
        - bearerAuth: []
      parameters:
        - name: saleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Venda finalizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdvSaleResponse'
        '400':
          $ref: '#/components/responses/BusinessError'
        '403':
          $ref: '#/components/responses/BusinessError'
        '404':
          $ref: '#/components/responses/BusinessError'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /api/pdv/sales/{saleId}/discount:
    post:
      summary: Aplica desconto ao total da venda PDV.
      operationId: applyPdvSaleDiscount
      security:
        - bearerAuth: []
      parameters:
        - name: saleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDiscountRequest'
      responses:
        '200':
          description: Desconto aplicado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdvSaleResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/BusinessError'
        '404':
          $ref: '#/components/responses/BusinessError'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /api/pdv/sales/{saleId}/cancel:
    post:
      summary: Cancela uma venda PDV em aberto.
      operationId: cancelPdvSale
      security:
        - bearerAuth: []
      parameters:
        - name: saleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelSaleRequest'
      responses:
        '200':
          description: Venda cancelada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdvSaleResponse'
        '403':
          $ref: '#/components/responses/BusinessError'
        '404':
          $ref: '#/components/responses/BusinessError'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /api/pdv/sales/{saleId}/print:
    post:
      summary: Solicita a reimpressão do comprovante de venda.
      operationId: reprintPdvSaleReceipt
      security:
        - bearerAuth: []
      parameters:
        - name: saleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Impressão enviada com sucesso.
        '403':
          $ref: '#/components/responses/BusinessError'
        '404':
          $ref: '#/components/responses/BusinessError'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /api/kds/tickets:
    get:
      summary: Lista tickets de preparo ativos no KDS.
      operationId: getKdsTickets
      security:
        - bearerAuth: []
      parameters:
        - name: sector
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/PreparationSector'
      responses:
        '200':
          description: Tickets encontrados.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KdsTicketResponse'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /api/kds/tickets/{id}/start:
    patch:
      summary: Marca início do preparo de um ticket.
      operationId: startKdsTicket
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ticket atualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KdsTicketResponse'
        '404':
          $ref: '#/components/responses/BusinessError'
        default:
          $ref: '#/components/responses/UnexpectedError'
...
  /api/kds/tickets/{id}/finish:
    patch:
      summary: Marca ticket como finalizado/entregue ao cliente.
      operationId: finishKdsTicket
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ticket atualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KdsTicketResponse'
        '404':
          $ref: '#/components/responses/BusinessError'
        default:
          $ref: '#/components/responses/UnexpectedError'
...
  /api/kds/orders/{id}/finish:
    patch:
      summary: Finaliza todos os tickets vinculados a um pedido.
      operationId: finishKdsOrder
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Tickets do pedido finalizados.
        '404':
          $ref: '#/components/responses/BusinessError'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /api/orders/{orderId}/payments:
    post:
      summary: Registra pagamento e finaliza venda local.
      operationId: createOrderPayment
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '201':
          description: Pagamento registrado e venda finalizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/BusinessError'
        '404':
          $ref: '#/components/responses/BusinessError'
        '409':
          $ref: '#/components/responses/BusinessError'
        default:
          $ref: '#/components/responses/UnexpectedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    ValidationError:
      description: Requisicao invalida por erro de validacao.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    BusinessError:
      description: Regra de negocio impediu a operacao.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    UnexpectedError:
      description: Erro inesperado sem vazamento de detalhes internos.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ApiError:
      type: object
      required:
        - code
        - message
        - status
        - path
        - timestamp
        - validationErrors
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Campos invalidos na requisicao.
        status:
          type: integer
          example: 400
        path:
          type: string
          example: /api/ping
        timestamp:
          type: string
          format: date-time
        validationErrors:
          type: array
          items:
            $ref: '#/components/schemas/ApiFieldError'
    ApiFieldError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          example: name
        message:
          type: string
          example: must not be blank
...
    CreatePdvSaleItemRequest:
      type: object
      required:
        - productId
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: number
          format: decimal
          minimum: 0.001
          example: 1
        observation:
          type: string
    PatchPdvSaleItemRequest:
      type: object
      required:
        - quantity
      properties:
        quantity:
          type: number
          format: decimal
          minimum: 0.001
          example: 2
    CreateDiscountRequest:
      type: object
      required:
        - amount
      properties:
        amount:
          type: number
          format: double
          minimum: 0
          example: 5.50
    CancelSaleRequest:
      type: object
      required:
        - reason
      properties:
        reason:
          type: string
          example: Cliente desistiu da compra
    PdvSaleResponse:
      type: object
      required:
        - id
        - origin
        - status
        - paymentStatus
        - deliveryType
        - subtotal
        - discountAmount
        - deliveryFee
        - totalAmount
        - items
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        orderNumber:
          type: integer
          format: int64
        origin:
          $ref: '#/components/schemas/OrderOrigin'
        status:
          $ref: '#/components/schemas/OrderStatus'
        paymentStatus:
          $ref: '#/components/schemas/PaymentStatus'
        deliveryType:
          $ref: '#/components/schemas/DeliveryType'
        subtotal:
          type: number
          format: decimal
        discountAmount:
          type: number
          format: decimal
        deliveryFee:
          type: number
          format: decimal
        totalAmount:
          type: number
          format: decimal
        items:
          type: array
          items:
            $ref: '#/components/schemas/PdvSaleItemResponse'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PdvSaleItemResponse:
      type: object
      required:
        - id
        - orderId
        - productNameSnapshot
        - quantity
        - unitPrice
        - totalPrice
        - status
        - preparationSector
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        productNameSnapshot:
          type: string
        quantity:
          type: number
          format: decimal
        unitPrice:
          type: number
          format: decimal
        totalPrice:
          type: number
          format: decimal
        observation:
          type: string
        status:
          $ref: '#/components/schemas/OrderItemStatus'
        preparationSector:
          $ref: '#/components/schemas/PreparationSector'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    KdsTicketResponse:
      type: object
      required:
        - id
        - orderId
        - ticketNumber
        - origin
        - sector
        - status
        - createdAt
        - items
      properties:
        id:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        ticketNumber:
          type: integer
          format: int64
        origin:
          $ref: '#/components/schemas/OrderOrigin'
        sector:
          $ref: '#/components/schemas/PreparationSector'
        status:
          $ref: '#/components/schemas/KdsTicketStatus'
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        readyAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/KdsTicketItemResponse'
    KdsTicketItemResponse:
      type: object
      required:
        - id
        - productName
        - quantity
        - status
      properties:
        id:
          type: string
          format: uuid
        productName:
          type: string
        quantity:
          type: number
        observation:
          type: string
        status:
          $ref: '#/components/schemas/KdsTicketStatus'
    KdsTicketStatus:
      type: string
      enum: [WAITING, IN_PREPARATION, READY, FINISHED, CANCELED]
      example: WAITING
    CreatePaymentRequest:
      type: object
      required:
        - method
        - amount
      properties:
        method:
          $ref: '#/components/schemas/PaymentMethod'
        amount:
          type: number
          format: decimal
          minimum: 0.01
        transactionId:
          type: string
        gateway:
          type: string
    PaymentResponse:
      type: object
      required:
        - id
        - orderId
        - method
        - status
        - amount
        - orderStatus
        - orderPaymentStatus
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        method:
          $ref: '#/components/schemas/PaymentMethod'
        status:
          $ref: '#/components/schemas/PaymentStatus'
        amount:
          type: number
          format: decimal
        transactionId:
          type: string
        gateway:
          type: string
        paidAt:
          type: string
          format: date-time
        canceledAt:
          type: string
          format: date-time
        orderStatus:
          $ref: '#/components/schemas/OrderStatus'
        orderPaymentStatus:
          $ref: '#/components/schemas/PaymentStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    OrderOrigin:
      type: string
      enum:
        - PDV
        - SITE
        - WHATSAPP
        - ADMIN
        - MANUAL
        - INTEGRATION
    OrderStatus:
      type: string
      enum:
        - CREATED
        - PAYMENT_PENDING
        - PAID
        - SENT_TO_STORE
        - RECEIVED_BY_STORE
        - ACCEPTED
        - IN_PREPARATION
        - READY
        - OUT_FOR_DELIVERY
        - DELIVERED
        - FINISHED
        - CANCELED
    PaymentStatus:
      type: string
      enum:
        - PENDING
        - AUTHORIZED
        - PAID
        - REFUSED
        - CANCELED
        - REFUNDED
        - EXPIRED
    DeliveryType:
      type: string
      enum:
        - COUNTER
        - PICKUP
        - DELIVERY
        - LOCAL_CONSUMPTION
    OrderItemStatus:
      type: string
      enum:
        - CREATED
        - WAITING_PREPARATION
        - IN_PREPARATION
        - READY
        - DELIVERED
        - CANCELED
    PreparationSector:
      type: string
      enum:
        - BALCAO
        - CHAPA
        - BEBIDAS
        - CONFEITARIA
        - EXPEDICAO
        - DELIVERY
        - SEM_PREPARO
...
    PaymentMethod:
      type: string
      enum:
        - CASH
        - PIX
        - CREDIT_CARD
        - DEBIT_CARD
        - ONLINE_PIX
        - ONLINE_CREDIT_CARD
        - ONLINE_DEBIT_CARD
        - MEAL_VOUCHER
        - MIXED
...